### 2021-07-01 #3443 @whedon
[PRE REVIEW]: Reel1.0 - A visualization tool for evaluating powder diffraction refinements
**Submitting author:** @fgjorup (<a href="http://orcid.org/0000-0003-3902-0247">Frederik Holm Gjørup</a>)
**Repository:** <a href="https://github.com/fgjorup/Reel/" target ="_blank">https://github.com/fgjorup/Reel/</a>
**Version:** v.1.1.1
**Reviewers:** @cmbiwer, mikapfl
