+++ .github/workflows/draft-pdf.yml
@@ -1,9 +1,23 @@
-- name: Open Journals PDF Generator
-  # You may pin to the exact commit or the version.
-  # uses: openjournals/openjournals-draft-action@a994a876f36e25076174d9ed133e01b1f1b9fcf3
-  uses: openjournals/openjournals-draft-action@v.1.0
-  with:
-    # Target journal, must be either "joss" or "jose"
-    journal: # optional, default is joss
-    # Path to the main paper Markdown file, relative to the project's root directory.
-    paper-path: # optional, default is paper.md
+on: [push]
+
+jobs:
+  paper:
+    runs-on: ubuntu-latest
+    name: Paper Draft
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+      - name: Build draft PDF
+        uses: openjournals/openjournals-draft-action@master
+        with:
+          journal: joss
+          # This should be the path to the paper within your repo.
+          paper-path: paper.md
+      - name: Upload
+        uses: actions/upload-artifact@v1
+        with:
+          name: paper
+          # This is the output path where Pandoc will write the compiled
+          # PDF. Note, this should be the same directory as the input
+          # paper.md
+          path: paper.pdf
