### 2021-07-16 #3443 @rkurchin
@cmbiwer, @mikapfl, would you be willing to review this submission for JOSS?

We carry out our checklist-driven reviews here in GitHub issues according to [these guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). You can find out more about this submission in the first comment in this issue. Note that our editorial robot whedon is having trouble compiling the PDF right now, but if you hop over to [the submission branch of their repo](https://github.com/fgjorup/Reel/tree/JOSS-submission) you can see the manuscript as well.
