### 2021-07-19 #3443 @mikapfl
Hi,

I haven't done any Rietveld refinement myself, so I can't really review the scientific usefulness of the software and how it relates to the wider ecosystem of powder diffraction analysis software. If that is fine with you, I would be happy to do a review. (-:

Cheers,

Mika
