### 2021-07-23 #3443 @rkurchin
@marcocamma, would you be able/willing to review this submission to JOSS?

We carry out our checklist-driven reviews here in GitHub issues according to these guidelines. You can find out more about this submission in the first comment in this issue. Note that our editorial robot whedon is having trouble compiling the PDF right now, but if you hop over to the submission branch of their repo you can see the manuscript as well.
