### 2021-07-28 #3443 @rkurchin
> @fgjorup, could you please correct the ORCID issue referenced above? Currently, your author/ORCID list is formatted thusly:
>
> ```
>   - name: Frederik H. Gjørup , Mathias Mørch, & Mogens Christensen
>     orcid: 0000-0003-3902-0247, 0000-0002-8022-5538, 0000-0001-6805-1232
>     affiliation: "1"
> ```
>
> I believe replacing it with the following should fix it:
>
> ```
>   - name: Frederik H. Gjørup
>     orcid: 0000-0003-3902-0247
>     affiliation: "1"
>
>   - name: Mathias Mørch
>     orcid: 0000-0002-8022-5538
>     affiliation: "1"
>
>   - name: Mogens Christensen
>     orcid: 0000-0001-6805-1232
>     affiliation: "1"
> ```

@fgjorup, we should be close to having two reviewers and being ready to move this to the review stage, if you could resolve this so that we can build a proof, that would be great!
