### 2021-07-29 #3546 @whedon
[REVIEW]: Reel1.0 - A visualization tool for evaluating powder diffraction refinements
**Submitting author:** @fgjorup (<a href="http://orcid.org/0000-0003-3902-0247">Frederik Holm Gjørup</a>)
**Repository:** <a href="https://github.com/fgjorup/Reel/" target ="_blank">https://github.com/fgjorup/Reel/</a>
**Version:** v1.2.0
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.16817929" target="_blank">10.6084/m9.figshare.16817929</a>
