### 2021-07-29 fgjorup/Reel#3 @mikapfl
LICENSE and README.md contradict each other
Hi,

according to the LICENSE file, Reel is published under the MIT license; however, the README.md file states:
```
The software provided here is intended only for internal use at Department of Chemistry, Aarhus University, and should not be distributed outside of the institute without the consent of the author.
```
Which clearly poses additional restrictions not imposed by the MIT license. I guess it is just a leftover from earlier? Preferably, you would just point to the license file with a statement such as "Reel is published under the MIT licence, details can be found in the LICENSE file." or similar.

Cheers,

Mika
