### 2021-07-29 #3546 @mikapfl
@fgjorup Could you describe how your co-authors contributed to the work? The git history doesn't show contributions from other people than you, but git histories of course only show one side of a larger work.
