### 2021-08-07 #3546 @fgjorup
> @fgjorup Could you describe how your co-authors contributed to the work? The git history doesn't show contributions from other people than you, but git histories of course only show one side of a larger work.

@mikapfl of course.

Professor Mogens Christensen is the supervisor of the project and has primarily contributed with scientific background, funding, and proof-reading of the paper.

Mathias Mørch has taken part in designing the layout of the interface and has been the primary tester of the program. He has also taken part in defining the format of the _.xyy_ input files and has provided the test files available in the repository.

Naturally, the above contributions have not been logged by Github.

I hope that answers your question, otherwise I will gladly elaborate.
