### 2021-08-07 #3546 @fgjorup
> @fgjorup, reminder to answer @mikapfl's question when you get a chance, as well as the issue he opened on your repository: [fgjorup/Reel#3](https://github.com/fgjorup/Reel/issues/3)
>
> @cmbiwer, please reach out if you have any questions about getting your review started.

@rkurchin, sorry for the delay, I am finally back in the office from my vacation. Please let me know if there is anything else I need to address.
