### 2021-08-12 #3546 @mikapfl
@fgjorup Could you provide a set of example files which I could use to run the program? I unfortunately don't have a set of XRD data files which would lend itself to an analysis with Reel.
