### 2021-08-16 #3546 @fgjorup
@mikapfl  There should already be a test_files folder in the Joss-submission repo, I hope those will do the trick.
