### 2021-08-16 #3546 @cmbiwer
I have finished taking a look at the software and can type up the checklist and my comments about the functionality, documentation, paper, and software sometime this week.

Though, there were a few crashes/errors I encountered. Either myself or @mikapfl have already made issues on the GitHub repo. I'd suggest those are address (i.e. issues #4 and #6).
