### 2021-08-27 2dca48d
Change of default path for openFiles method
Changed from os.environ to os.getenv. The open files dialog defaults to the user profile on windows and to the current directory on mac/linux.
