+++ Reel1.0/Refinement_evaluator_ver1.0.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Last update: 22/06/2021
+Last update: 27/08/2021
 Frederik H. Gjørup
 """
 
