### 2021-09-03 #3546 @cmbiwer
> ## Review checklist for @cmbiwer
> ### Conflict of interest
> * [X]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
> * [X]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [X]  **Repository:** Is the source code for this software available at the [repository url](https://github.com/fgjorup/Reel/)?
> * [X]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
> * [X]  **Contribution and authorship:** Has the submitting author (@fgjorup) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> * [X]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> ### Functionality
> * [X]  **Installation:** Does installation proceed as outlined in the documentation?
> * [X]  **Functionality:** Have the functional claims of the software been confirmed?
> * [X]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ### Documentation
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [X]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> * [X]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> * [X]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?
> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> ### Software paper
> * [X]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> * [ ]  **A statement of need:** Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
> * [X]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
> * [X]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?

I've filled out the checklist above. Overall, I believe the software fills a need and it is easy to imagine the software to be useful to scientists. Several other projects have developed or are developing similar 2D heatmap visualizations in some fashion. So I would say the functionality is recognized as an important development by the community already. I liked the user interface to Reel and I would say the author has created a nice-looking software product. However, this project is still missing some key items that are explicitly outlined in the checklist. Notably, there is no no community guidelines and the paper could give a bit clearer picture of Reel to the reader.

There is no community guidelines for contribution. The lack of community guidelines and no documentation of the code itself or inline comments will probably not invite much open-source contributions.

I would say the JOSS paper manuscript itself doesn't give a non-familiar reader a clear image of what Reel does. I think the text in the summary is very nice. However, an image of the GUI (like in their documentation) would give a reader of the actual paper a much better idea what Reel is in its statement of need. There also isn't a discussion of example usage or what real problems could be solved (for example identifying where a phase change happens, etc.). The authors mention refinement programs but don't really state much about the current state of visualization in the field which is the intent of Reel.

The instructions do show the users to manually load up the software and verify its functionality, so it does meet the JOSS requirements. However, there were a number of issues already and perhaps some kind of automated testing would be beneficial to maintaining this project.

I'd say Reel is a nice program, but its still missing some of the review checklist items. The author has already address several issues in the repo, perhaps they could address these final missing items as well.
