### 2021-09-07 8e9a4fa
Update README.md
Fixed table format and added section for contact info
