### 2021-09-13 #3546 @fgjorup
Hi @rkurchin  and @cmbiwer
I have added a line about contact on contribution in the README.md, I hope that satisfies the requirements.

Regarding the manuscript, I would be happy to include an image of the GUI, perhaps with a comparison to other common visualization tools in the community. I will also expand a bit on the statement of need, how I expect the community to use the software, and how it compares to what is already available.

Should I already now revise the manuscript, or should I wait for feedback from @mikapfl ?
