### 2021-09-13 #3546 @mikapfl
Hi,

regarding the manuscript, I think the summary is already good and for me there is no need to replicate things already found in the docs. However, if I understand correctly, Reel is rather one building block in a "pipeline" - while it is possible to view raw measurements, it really shines when viewing refinement results. In the docs, there is a list of supported file formats, maybe you could explain in the paper the ecosystem more, in particular which refinement software is "supported" by Reel and which hole Reel fills in the whole pipeline.

I will write a bit more regarding the other (non-paper) parts of the checklist and/or open more issues later.

Cheers,

Mika
