### 2021-09-13 #3546 @mikapfl
Regarding your changes in the `JOSS-submission` branch, do you intend to merge this into master? At the moment, people just landing on Reel's github presence don't get the various enhancements you already implemented.
