### 2021-09-15 fgjorup/Reel#11 @mikapfl
AttributeError: module 'PyQt5.QtCore' has no attribute 'Signal'
Hi,

after installing Reel (regardless if under windows or linux), in python 3.8 with the latest dependencies as stated in requirements.txt, I get the following error when starting Reel:

```python
Opening Reel 1.0...
Traceback (most recent call last):
  File "Refinement_evaluator_ver1.0.py", line 28, in <module>
    from _lib.ReelPlotWidgets import MultiImageWidget, PlotPatternWidget, PlotSliceWidget, PlotParametersWidget
  File "C:\Users\Hein Mück\PycharmProjects\Reel\Reel1.0\_lib\ReelPlotWidgets.py", line 30, in <module>
    class MultiImageWidget(pg.GraphicsLayoutWidget):
  File "C:\Users\Hein Mück\PycharmProjects\Reel\Reel1.0\_lib\ReelPlotWidgets.py", line 32, in MultiImageWidget
    sigHLineDragged = QtCore.Signal(object)
AttributeError: module 'PyQt5.QtCore' has no attribute 'Signal'
Drücken Sie eine beliebige Taste . . .
```

It seems I can solve this by mokey-patching PyQt5.QtCore at line 24 of `ReelPlotWidgets.py`, but that is a very strange and hackish solution:

```diff
+ QtCore.Signal = QtCore.pyqtSignal
```

The versions of libraries installed in my python are:

Package | Version
-- | -- | --
Pillow | 8.3.2
**PyQt5** | 5.15.4
**PyQt5-Qt5** | 5.15.2
PyQt5-sip | 12.9.0
cycler | 0.10.0
kiwisolver | 1.3.2
matplotlib | 3.4.3
numpy | 1.21.2
pip | 21.1.2
pyparsing | 2.4.7
**pyqtgraph** | 0.12.2
python-dateutil | 2.8.2
scipy | 1.7.1
setuptools | 57.0.0
six | 1.16.0
wheel | 0.36.2

I used virtualenv to create the python environment.

Cheers,

Mika
