### 2021-09-17 #3546 @mikapfl
Hi,

I've updated my checklist and reported some installation problems that should be rather straight-forward to fix.

I think Reel 1.0 is a fairly intuitively usable software which can be useful for analysis of preliminary results at the beamline as well as later on in the refinement process. However, I think to make Reel a solid foundation for future updates and maintenance, it needs automated tests. I don't agree with @cmbiwer that the instructions provided in the documentation fulfil the criteria for either automated tests or even clearly described steps for manual testing. A key functionality of Reel is that, according to the documentation, it is possible to analyse data supplied in 6 different formats, but at the moment, only the Reel custom format can be tested using the given instructions and data. Even there, expected test results are not clearly defined, so that more subtle errors can't be detected. As proper reading of file formats is often a source of errors, but also crucial for users, I think that testing should be done (preferably automatically) using all supported file formats. This is even possible with standard python tools for unit testing by testing the reading functions directly, so that more advanced testing tools for graphical user interfaces are not necessary.

Additionally, either automated tests or clear descriptions how to perform manual tests should be provided for testing Reel in a defined python environment. Some libraries Reel uses are under quite active development, and issues like https://github.com/fgjorup/Reel/issues/11 can be caught and fixed in time using testing strategies with clearly defined python environments.

Cheers,

Mika
