### 2021-09-24 fgjorup/Reel#11 @fgjorup
That is a very odd behavior. Looking at the documentation for `PyQt5.QtCore`, I can't figure out why the issue arises, but I'll try replacing `PyQt5.QtCore.Signal` with `PyQt5.QtCore.pyqtSignal`, hopefully that will solve future issues.
