### 2021-09-27 fgjorup/Reel#11 @mikapfl
At least if I look in the [reference guide for PyQt5.QtCore](https://docs.huihoo.com/pyqt/PyQt5/QtCore.html), I don't see a `Signal` object there. I am no PyQt expert maybe newer versions of PyQt don't have the `Signal` object anymore or so?
