### 2021-10-06 4af9cf0
Update paper.md
Added illustration of the interface, an example flowchart, and a comparison to other visualization tools in the community.
Expanded on Statement of need and added a Statement of Field
