+++ paper.md
@@ -29,13 +29,25 @@ bibliography: paper.bib
 ---
 
 # Summary
-The ever-growing  community for parameter-resolved x-ray and neutron diffraction, spurred by the rapid improvements in both detectors and sources for large-scale facilities, gives rise to the need for a fast and efficient evaluation approach for the large quantities of data produced during such experiments. As diffraction scientists we need to be able to both visually and analytically compare our raw data and refined models in a consistent and user-friendly way. This is particularly true when refining in two dimensions, such as sequential or parametric refinements, where parameters such as time, temperature, field strength, pressure, etc. might be included in the models. Our proposed strategy focus on combined 1D and 2D visualizations (heatmaps) of the data, in order to qualitatively evaluate the observed, calculated, and residual data in parameter-space, with cross-comparison to key parameters.  
+The ever-growing  community for parameter-resolved x-ray and neutron diffraction, spurred by the rapid improvements in both detectors and sources for large-scale facilities, gives rise to the need for a fast and efficient evaluation approach for the large quantities of data produced during such experiments. As diffraction scientists we need to be able to both visually and analytically compare our raw data and refined models in a consistent and user-friendly way. This is particularly true when refining in two dimensions, such as sequential or parametric refinements, where parameters such as time, temperature, field strength, pressure, etc. might be included in the models. Our proposed strategy focus on combined 1D and 2D visualizations (heatmaps) of the data, in order to qualitatively evaluate the observed, calculated, and residual data in parameter-space, with cross-comparison to key parameters. We accomplish this with a single-window interface, illustrated in \autoref{fig:Fig. 1}, where 1D and 2D data are easily compared with the moveable lines (or *Reel cursors*). Several datasets can be opened at once to quickly compare models. Using the simple customized *.xyy* file format allows any number of user-defined parameters to be plotted, such as temperature, pressure, R-values, or mean intensity.  
 
-# Statement of need
+![Illustration of the single-window interface of Reel1.0. The interface is divided in three main sections: 2D surface plots (heatmaps), 1D diffraction pattern plot, and 1D parameter plot.\label{fig:Fig. 1}](figure_1.tiff)
+
+# Statement of Need
 
 `Reel1.0` is a Python based GUI, based on the `PyQt5` and `pyqtgraph` packages [@pyqtgraph]. The graphical interface is intended to make `Reel1.0` appealing to a broad audience (within the community), even for users with limited programming knowledge. User-friendliness is a key feature, as `Reel1.0` is a visualization tool, intended to be used on combination with other refinement software, such as TOPAS[@topas] or FullProf[@fullprof], and with room for expansion in the future.
 
-`Reel1.0` is intended for users working with large x-ray and neutron diffraction datasets, such as in-situ and operando studies, across several scientific fields (chemistry, physics, materials science). The visualization tools provided by the `pyqtgraph` libraries allow multiple datasets of several hundred patterns to be evaluated at once, without compromising the stability of the program. The simplistic interface, efficiency, and user-friendliness of `Reel1.0` will allow the powder diffraction community to pursue increasingly advanced parameter-resolved experiments and the modeling of these.
[…]
