+++ paper.bib
@@ -16,6 +16,24 @@
     doi = {10.1016/0921-4526(93)90108-I},
 }
 
+@article{maud,
+  	Author = {{Lutterotti}, L. and {Bortolotti}, M. and {Ischia}, G. and {Lonardelli}, I and {Wenk}, H.-R.},
+  	Journal = {Zeitschrift Fur Kristallographie},
+  	Title = {{Rietveld texture analysis from diffraction images}},
+  	Year = 2007,
+    Volume = 26,
+    doi = {10.1524/9783486992540-020},
+}
+
+@article{nunez2018,
+  	Author = {{Nuñez}, J.R. and {Anderton}, C.R. and {Renslow}, R.S.},
+  	Journal = {Plos One},
+  	Title = {{Optimizing colormaps with consideration for color vision deficiency to enable accurate interpretation of scientific data}},
+  	Year = 2018,
+    Volume = 13,
+    doi = {10.1371/journal.pone.0199239},
+}
+
 @misc{pyqtgraph,
   	Author = {{Campagnola}, L.},
   	Journal = {Physica B},
