+++ paper.md
@@ -40,14 +40,14 @@ The ever-growing  community for parameter-resolved x-ray and neutron diffraction
 `Reel1.0` is intended for users working with large x-ray and neutron diffraction datasets, such as in-situ and operando studies, across several scientific fields (chemistry, physics, materials science). The visualization tools provided by the `pyqtgraph` libraries allow multiple datasets of several hundred patterns to be evaluated at once, without compromising the stability of the program. The simplistic interface, efficiency, and user-friendliness of `Reel1.0` will allow the powder diffraction community to pursue increasingly advanced parameter-resolved experiments and the modeling of these.  
 `Reel1.0` can be used at several steps in the diffraction data processing, both for evaluating raw data and for evaluating and comparing refinement models. A suggested workflow for diffraction data evaluation using Reel1.0 is outlined in \autoref{fig:figure_2}. During data collection, the raw data can be evaluated to ensure that an appropriate data quality is acquired. This includes evaluating parameters such as the signal-to-noise ratio, signal-to-background ratio, time-resolution, and angular-resolution. During initial assessment, a more thorough pre-modeling evaluation can be performed. Here, the now collected data is evaluated, in order to gather information needed for the modeling. This includes phase transition, impurity formation, changes in background, mean scattered intensity, and external parameters (when available). The initial assessment also serves as a quick quality-check, before the more time-consuming refinement is performed. The refinement is performed in an external software, e.g. FullProf or TOPAS, and the resulting output files are evaluated in `Reel1.0`. The modeling evaluation is particularly useful for direct visual assessment, both of 1D and 2D plots, but also for evaluating R-values and user-provided parameters (only for the *.xyy* format). The 2D residual surface plot is an exceptional tool for identifying systematic deviations in the model, such as asymmetry, diffuse scattering, or incorrect background modeling.  
 
-![Example of a data processing flowchart using Reel1.0, along with examples of the associated data file formats. The formats .dat and .prf are based on standard FullProf formats, while the .xyy format is a custom file format, defined in the Ree1.0 Quick Guide.\label{fig:figure_2}](figure_2.tiff)
+![Example of a data processing flowchart using Reel1.0, along with examples of the associated data file formats. The formats .dat and .prf are based on standard FullProf formats, while the .xyy format is a custom file format, defined in the Ree1.0 Quick Guide.\label{fig:figure_2}](figure_2.tif)
 
 # Statement of Field
 
 To our knowledge, no other software in the community provide an as easy and fast evaluation of both 1D and 2D diffraction patterns. While refinement software, such as FullProf Suite, MAUD[@maud], or TOPAS come with visualization tools for both 1D and 2D patterns, they lack the ability to combine the two. Examples of the visualization tools in FullProf Suite, MAUD, and TOPAS are illustrated in \autoref{fig:figure_3}.	
 The aesthetics of the visualizations are of course subjective, however, it has been shown that certain colormaps, such as “rainbow” maps, can be misleading and in the worst case, result in scientifically wrong conclusions.[@nunez2018] As such, one should strive for using perceptually linear colormaps. Reel1.0 uses the perceptually linear and color vision deficiency friendly `cividis` colormap as default, but any colormap contained in the `Matplotlib` package is available.  
 
-![Examples of other visualization tools in the powder diffraction community. a) A 1D diffraction pattern as visualized in FullProf Suites WinPlotr. b) 2D observed-calculated heatmaps for 72 patterns as visualized in MAUD. c) Observed, calculated, and residual waterfall plot as visualized in TOPAS.\label{fig:figure_3}](figure_3.tiff)
[…]
