### 2021-10-06 #3546 @fgjorup
Thank you @rkurchin , we had a very productive trip to Sweden, where we managed to do some very exciting science at the new DanMAX beamline.

Inspired by both @cmbiwer  and @mikapfl, I have updated the paper.md, so it now includes a figure of the interface, a flowchart-figure, and a figure comparing other visualization option in the field. I have also expanded on the _Statement of Need_ and added a _Statement of Field_.

I will try to get around to addressing the remaining issues and look in to testing, as @mikapfl points out.
