### 2021-10-07 42cd7e6
Changed QtCore.Signal to QtCore.pywtSignal - Issue #11
