### 2021-10-07 d477847
Update Refinement_evaluator_ver1.0.py
- Removed .par file as a suggested file format, but retained the underlying code.
- Added "debug" option. Run `Refinement_evaluator_ver1.0.py` with _-debug_ as argument to open all file formats in __test_files_.
