### 2021-10-07 ebaa653
Add debug information
