+++ paper.md
@@ -35,7 +35,7 @@ The ever-growing  community for parameter-resolved x-ray and neutron diffraction
 
 # Statement of Need
 
-`Reel1.0` is a Python based GUI, based on the `PyQt5` and `pyqtgraph` packages [@pyqtgraph]. The graphical interface is intended to make `Reel1.0` appealing to a broad audience (within the community), even for users with limited programming knowledge. User-friendliness is a key feature, as `Reel1.0` is a visualization tool, intended to be used on combination with other refinement software, such as TOPAS[@topas] or FullProf[@fullprof], and with room for expansion in the future.
+`Reel1.0` is a Python based GUI, based on the `PyQt5` and `pyqtgraph` packages [@pyqtgraph]. The graphical interface is intended to make `Reel1.0` appealing to a broad audience (within the community), even for users with limited programming knowledge. User-friendliness is a key feature, as `Reel1.0` is a visualization tool, intended to be used in combination with other refinement software, such as TOPAS[@topas] or FullProf[@fullprof], and with room for expansion in the future.
 
 `Reel1.0` is intended for users working with large x-ray and neutron diffraction datasets, such as in-situ and operando studies, across several scientific fields (chemistry, physics, materials science). The visualization tools provided by the `pyqtgraph` libraries allow multiple datasets of several hundred patterns to be evaluated at once, without compromising the stability of the program. The simplistic interface, efficiency, and user-friendliness of `Reel1.0` will allow the powder diffraction community to pursue increasingly advanced parameter-resolved experiments and the modeling of these.  
 `Reel1.0` can be used at several steps in the diffraction data processing, both for evaluating raw data and for evaluating and comparing refinement models. A suggested workflow for diffraction data evaluation using Reel1.0 is outlined in \autoref{fig:figure_2}. During data collection, the raw data can be evaluated to ensure that an appropriate data quality is acquired. This includes evaluating parameters such as the signal-to-noise ratio, signal-to-background ratio, time-resolution, and angular-resolution. During initial assessment, a more thorough pre-modeling evaluation can be performed. Here, the now collected data is evaluated, in order to gather information needed for the modeling. This includes phase transition, impurity formation, changes in background, mean scattered intensity, and external parameters (when available). The initial assessment also serves as a quick quality-check, before the more time-consuming refinement is performed. The refinement is performed in an external software, e.g. FullProf or TOPAS, and the resulting output files are evaluated in `Reel1.0`. The modeling evaluation is particularly useful for direct visual assessment, both of 1D and 2D plots, but also for evaluating R-values and user-provided parameters (only for the *.xyy* format). The 2D residual surface plot is an exceptional tool for identifying systematic deviations in the model, such as asymmetry, diffuse scattering, or incorrect background modeling.  
