### 2021-10-14 0c46ce2
Update paper.md
https://github.com/openjournals/joss-reviews/issues/3546#issuecomment-943411197
