### 2021-10-14 85fcdd1
Update paper.bib
https://github.com/openjournals/joss-reviews/issues/3546#issuecomment-943411197
