### 2021-10-14 #3546 @fgjorup
Great news @rkurchin ! My colleague, @moerch, should have taken care of the "editorial feedback", so the manuscript should be ready. Regarding merging, which makes the most sense? I would like the archived and tagged release to include the changes in the JOSS branch, so would you prefer I merged the branch before acceptance?
